Refactor common methods into parent class DataAssistantResult
